refactor(frontend): rename Activity to AllTransactions
