+++ src/frontend/src/lib/components/transactions/AllTransactions.svelte
