+++ src/frontend/src/routes/(app)/activity/+page.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
 	import MultipleListeners from '$lib/components/core/MultipleListeners.svelte';
-	import Activity from '$lib/components/transactions/Activity.svelte';
+	import AllTransactions from '$lib/components/transactions/AllTransactions.svelte';
 	import { enabledNetworkTokens } from '$lib/derived/network-tokens.derived';
 </script>
 
 <MultipleListeners tokens={$enabledNetworkTokens}>
-	<Activity />
+	<AllTransactions />
 </MultipleListeners>
