+++ src/frontend/src/tests/lib/components/transactions/AllTransactions.spec.ts
@@ -1,11 +1,11 @@
-import Activity from '$lib/components/transactions/Activity.svelte';
+import AllTransactions from '$lib/components/transactions/AllTransactions.svelte';
 import en from '$tests/mocks/i18n.mock';
 import { assertNonNullish } from '@dfinity/utils';
 import { render } from '@testing-library/svelte';
 
 describe('Activity', () => {
 	it('renders the title', () => {
-		const { container } = render(Activity);
+		const { container } = render(AllTransactions);
 
 		const title: HTMLHeadingElement | null = container.querySelector('h1');
 
@@ -16,7 +16,7 @@ describe('Activity', () => {
 	});
 
 	it('renders the transactions list', () => {
-		const { getByText } = render(Activity);
+		const { getByText } = render(AllTransactions);
 
 		expect(getByText(en.transactions.text.transaction_history)).toBeInTheDocument();
 	});
